[refactor] #55 music 데이터 구조 변경
