+++ core/ui/src/main/java/com/example/ui/model/BookmarkedMusic.kt
@@ -1,0 +1,6 @@
+package com.example.ui.model
+
+data class BookmarkedMusic(
+    val postId: Long,
+    val music: Music
+)
