+++ core/ui/src/main/java/com/example/ui/model/Music.kt
@@ -1,0 +1,8 @@
+package com.example.ui.model
+
+data class Music(
+    val trackId: String,
+    val musicTitle: String,
+    val artistName: String,
+    val thumbnailUrl: String?,
+)
