+++ core/ui/src/main/java/com/example/ui/model/RegisteredMusic.kt
@@ -1,0 +1,7 @@
+package com.example.ui.model
+
+data class RegisteredMusic(
+    val postId: Long,
+    val comment: String,
+    val music: Music
+)
